pyproject.toml
[tool.poetry]
name = "nycu-data-science-2024-makeup-transfer"
version = "0.1.0"
description = ""
authors = ["KeithLin724 <keithlin724@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
pandas = "^2.2.2"
rich = "^13.7.1"
black = "^24.4.2"
torch = "^2.3.1"
torchvision = "^0.18.1"
torchaudio = "^2.3.1"
easydict = "^1.13"
opencv-python = "^4.10.0.82"
tqdm = "^4.66.4"
matplotlib = "^3.9.0"
pillow = "^10.3.0"
pydantic = "^2.7.3"
tensorboardx = "^2.6.2.2"
tensorboard = "^2.17.0"
albumentations = "^1.4.8"
tabulate = "^0.9.0"
click = "^8.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
